[40] - Folders refactor
