+++ app/Infrastructure/Controllers/WalletCryptocurrenciesController.php
@@ -3,7 +3,7 @@
 namespace App\Infrastructure\Controllers;
 
 use App\Application\Services\WalletCryptocurrenciesService;
-use App\Infrastructure\Validators\WalletIdValidator;
+use App\Infrastructure\Controllers\Validators\WalletIdValidator;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Response;
 use Illuminate\Routing\Controller as BaseController;
