+++ bootstrap/app.php
@@ -33,7 +33,7 @@ $app->singleton(
 
 $app->singleton(
     Illuminate\Contracts\Console\Kernel::class,
-    \App\Infrastructure\Console\Kernel::class
+    App\Infrastructure\Console\Kernel::class
 );
 
 $app->singleton(
