+++ tests/app/Infrastructure/Controllers/SellCoinControllerTest.php
