+++ tests/app/Infrastructure/Controllers/WalletBalanceControllerTest.php
