+++ tests/app/Infrastructure/Controllers/WalletCryptocurrenciesControllerTest.php
